+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Generate models from Figma's REST API specification ([@donny-dont](https://github.com/donny-dont))
 - Update to v0.40.0 of REST API specification ([@donny-dont](https://github.com/donny-dont))
 - Add `Slot` support ([@donny-dont](https://github.com/donny-dont))
+- Expose additional error information from API ([@donny-dont](https://github.com/donny-dont))
 - Require `sdk: ^3.8.0`
 
 ## 7.5.0